+++ backend/jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/flowable/dto/FlowableTaskContextResp.java
@@ -19,6 +19,9 @@ public class FlowableTaskContextResp {
     @ApiModelProperty("process name")
     private String processName;
 
+    @ApiModelProperty("business key")
+    private String businessKey;
+
     @ApiModelProperty("record id")
     private String recordId;