+++ backend/jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/flowable/service/IFlowableProcessService.java
@@ -36,6 +36,8 @@ public interface IFlowableProcessService {
 
     FlowableTaskContextResp getTaskContext(String taskId);
 
+    FlowableTaskContextResp getProcessContext(String processInstanceId);
+
     List<FlowableProcessTraceResp> getProcessTrace(String processInstanceId);
 
     List<FlowableHistoricTaskResp> queryDoneTasks(FlowableTaskQueryReq req, String username);